Added flex-wrap to few divs
Added flex-wrap to the information div and the content div, so they wrap when browser is scaled down. There is still an issue where the recipie-information and the ingredients divs overlap, still looking for a way to prevent this as when I tried it with margin it will apply for all sizes, not only when its scaled down.